core, refactor: split the source code to simplify fibjs.cpp.
