+++ fibjs/fibjs.vcxproj
@@ -201,6 +201,7 @@
     <ClInclude Include="include\encoding.h" />
     <ClInclude Include="include\encoding_bson.h" />
     <ClInclude Include="include\encoding_iconv.h" />
+    <ClInclude Include="include\fibjs.h" />
     <ClInclude Include="include\ifs\Buffer.h" />
     <ClInclude Include="include\ifs\BufferedStream.h" />
     <ClInclude Include="include\ifs\Chain.h" />
@@ -365,6 +366,7 @@
     <ClCompile Include="src\base\date.cpp" />
     <ClCompile Include="src\base\date_cache.cpp" />
     <ClCompile Include="src\base\fibjs.cpp" />
+    <ClCompile Include="src\base\fibjs_main.cpp" />
     <ClCompile Include="src\base\fuck_sym.cpp" />
     <ClCompile Include="src\base\opt_tools.cpp" />
     <ClCompile Include="src\base\options.cpp" />
